* refactor: changed builtin types so they match with postgres pg_catalog types and be able to use their qualified names when registering dumpers/builders and when working with migration builder. * refactor: primary keys, foreign keys, unique constraints and indexes will be defined through decorators, as column order wasnt preserved when extracting them from field metadata. the only metadata that remains is default values and checks. * refactor: added in_schema decorator to declare the namespace the definition belongs to and be able to work with qualified names when migrating. * refactor: meta classes now live in the meta namespace. * feature: added migration builder and its test file test_migration_builder.py convenient to generate alter/create/drop sentences to apply on the database.

+++ test/test_app/sql/functions.sql
@@ -1,7 +1,7 @@
 
 
 CREATE OR REPLACE FUNCTION get_author_by_id(
-    p_author_id bigint
+    p_author_id int8
 ) RETURNS author AS $$
 DECLARE
     v_result author;
@@ -16,7 +16,7 @@ $$ LANGUAGE plpgsql;
 
 
 CREATE OR REPLACE FUNCTION get_post_by_id(
-    p_post_id bigint
+    p_post_id int8
 ) RETURNS post AS $$
 DECLARE
     v_result post;
